requirements.txt
aiogram~=3.14.0
SQLAlchemy~=2.0.36
psycopg2-binary~=2.9.10
python-dotenv~=1.0.1
